move prebuild detail to correct folder
